Upload final code for P3
may be buggy because I fried my feather at the start of demos and lost most recent version of code, so had to rewrite the run_dc() function. Lesson learned to make sure to remember to double check I've saved code when I move on to focus on physical fabrication.